change definition of variable $dir
During set of variable localcontent, we assume file configure.ac is located under /root/3dcoin, why search for this directory after that assumption? More with ubuntu 18.04, the builtin cd of bash exit with staus 1 "cd: too many arguments" if more than one directory is found. That is our case "https://3dctalk.net/topic/1141/implemented-now-the-function-of-autoupdate-for-masternodes-howto-install-from-official-script/4"